+++ middleware_api/lambda/checkpoints/update_checkpoint.py
@@ -6,8 +6,8 @@ from typing import Any, Dict
 
 from common.ddb_service.client import DynamoDbUtilsService
 from common.response import ok, bad_request, internal_server_error
-from common_tools import complete_multipart_upload
 from lib._types import CheckPoint, CheckPointStatus
+from lib.common_tools import complete_multipart_upload
 
 checkpoint_table = os.environ.get('CHECKPOINT_TABLE')
 bucket_name = os.environ.get('S3_BUCKET')
