+++ middleware_api/lambda/endpoints/list_endpoints.py
@@ -2,12 +2,12 @@ import logging
 import os
 
 import boto3
+from multi_users._types import PARTITION_KEYS, Role
 
-from _enums import EndpointStatus
+from lib._enums import EndpointStatus
 from _types import EndpointDeploymentJob
 from common.ddb_service.client import DynamoDbUtilsService
 from common.response import ok, bad_request
-from multi_users._types import PARTITION_KEYS, Role
 from multi_users.utils import get_user_roles, check_user_permissions, get_permissions_by_username
 
 sagemaker_endpoint_table = os.environ.get('DDB_ENDPOINT_DEPLOYMENT_TABLE_NAME')
