--- middleware_api/lambda/inference_v2/_enums.py
@@ -1,12 +1,0 @@
-from enum import Enum, unique
-
-
-@unique
-class EndpointStatus(Enum):
-    CREATING = "Creating"
-    IN_SERVICE = "InService"
-    DELETED = "Deleted"
-    DELETING = "Deleting"
-    FAILED = "Failed"
-    UPDATING = "Updating"
-    ROLLING_BACK = "RollingBack"
