+++ middleware_api/lambda/lib/_types.py
@@ -88,3 +88,142 @@ class TrainJob:
 class MultipartFileReq:
     filename: str
     parts_number: int
+
+
+from dataclasses import dataclass
+from typing import Optional, Any
+
+import boto3.dynamodb.types
+
+Default_Role = 'IT Operator'
+
+
+@dataclass
+class _PartitionKeys:
+    user = 'user'
+    role = 'role'
+    # permission = 'permission'
+
+
+PARTITION_KEYS = _PartitionKeys()
+
+
+@dataclass
+class BaseMultiUserEntity:
+    kind: str
+    sort_key: str
+    creator: str
+
+
+@dataclass
+class User(BaseMultiUserEntity):
+    password: bytes
+    roles: [str]
+    params: Optional[dict[str, Any]] = None
+
+    def __post_init__(self):
+        if type(self.password) in (boto3.dynamodb.types.Binary, boto3.dynamodb.types.BINARY):
+            self.password = self.password.value
+
+
+@dataclass
+class Role(BaseMultiUserEntity):
+    permissions: [str]
+    params: Optional[dict[str, Any]] = None
+
+# @dataclass
+# class Permission(BaseMultiUserEntity):
+#     name: str
+#     params: Optional[dict[str, Any]] = None
+
+from dataclasses import dataclass
+from enum import Enum
+from typing import Optional, Any
+
+
+class DataStatus(Enum):
+    Initialed = 'Initialed'
+    Enabled = 'Enabled'
+    Disabled = 'Disabled'
+
+
+@dataclass
+class DatasetItem:
+    dataset_name: str  # partition key (s3 key base, not include bucket name)
+    sort_key: str  # sorted key: timestamp_name
+    name: str  # data name (s3 object key)
+    type: str  # data type
+    data_status: DataStatus
+    params: Optional[dict[str, Any]] = None
+    allowed_roles_or_users: Optional[list[str]] = None
+
+    def __post_init__(self):
+        if type(self.data_status) == str:
+            self.data_status = DataStatus[self.data_status]
+
+    def get_s3_key(self):
+        return f'dataset/{self.dataset_name}/{self.name}'
+
+
+class DatasetStatus(Enum):
+    Initialed = 'Initialed'
+    Enabled = 'Enabled'
+    Disabled = 'Disabled'
+
+
+@dataclass
+class DatasetInfo:
+    dataset_name: str  # primary key
+    timestamp: float
+    dataset_status: DatasetStatus
+    params: Optional[dict[str, Any]] = None
+    allowed_roles_or_users: Optional[list[str]] = None
+
+    def __post_init__(self):
+        if type(self.dataset_status) == str:
+            self.dataset_status = DatasetStatus[self.dataset_status]
+
+    def get_s3_key(self):
+        return f'dataset/{self.dataset_name}'
+
+
+from dataclasses import dataclass
+from typing import Optional, Any, List
+
+
+@dataclass
+class InferenceJob:
+    InferenceJobId: str
+    startTime: str
+    status: str
+    taskType: str
+    owner_group_or_role: Optional[List[str]] = None
+    inference_info_name: Optional[Any] = None
+    image_names: Optional[Any] = None
+    sagemakerRaw: Optional[Any] = None
+    completeTime: Optional[Any] = None
+    params: Optional[dict[str, Any]] = None
+
+
+@dataclass
+class EndpointDeploymentJob:
+    EndpointDeploymentJobId: str
+    autoscaling: bool
+    max_instance_number: str
+    startTime: str
+    status: str = None  # deprecated, but can't remove, avoid unexpected keyword argument
+    current_instance_count: str = None
+    endTime: Optional[str] = None
+    endpoint_status: Optional[str] = None
+    endpoint_name: Optional[str] = None
+    error: Optional[str] = None
+    owner_group_or_role: Optional[List[str]] = None
+
+
+# a copy of aws_extensions.models.InvocationsRequest
+@dataclass
+class InvocationsRequest:
+    task: str
+    username: Optional[str]
+    param_s3: str
+    models: Optional[dict]
