+++ middleware_api/lambda/lib/common_tools.py
