+++ middleware_api/lambda/roles/delete_roles.py
@@ -7,7 +7,7 @@ import boto3
 
 from common.response import ok, bad_request
 # todo will be optimize later
-from multi_users._types import PARTITION_KEYS, Default_Role
+from lib._types import PARTITION_KEYS, Default_Role
 
 logger = logging.getLogger('delete_roles')
 logger.setLevel(logging.INFO)
